BOCM-205 - renames files for consistency
